requirements.txt
streamlit==1.31.1
langchain==0.1.9
python-dotenv==0.21.0
huggingface-hub==0.20.3
qdrant-client==1.8.0
pypdf==4.0.2
langchain-groq==0.1.3